refactor: move stats context down to DailyStatisticsScreen
